backend/requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.12.2
ciso8601==2.2.0
fastapi==0.95.0
lxml==4.9.2
pandas==1.5.2
psycopg2==2.9.6
psycopg2_binary==2.9.5
pydantic==1.10.2
pytest==7.2.1
python-dotenv==1.0.0
redis==4.4.0
uvicorn==0.21.1
